Create README.md
This is a voice recognition software build for Tangerine bank. Tangerine has very limited ATM'S/ABM's of their own and rely heavily on Scotia, 7 eleven and Exchange network ABM's for their customer's banking needs. Since they have no bank and branch of their own its super hard to find their ATM's in google maps. The voice search for the tangerine ABM's on google and siri doesnt recognize these ATM's and give in appropriate results.

The goal of this project is to build a system which takes location from the user and generates a voice output about the nearest bank atm.